Shiomi Sawa Interview.
I'm excited to share the interview we conducted with Shiomi Sawa, the CEO of Hi-5 and publisher of Bright Kingdom. This conversation took place back in January of this year. A big thank you to Dashie, the entire team, and all the contributors who made this possible. We reached out to Sawa via email, and he kindly provided insights into his experience working with OnsOn Soft and what goes on behind the scenes in publishing.